feat(win): reuse common code
