+++ winio-ui-winui/src/ui/filebox.rs
@@ -1,43 +1,15 @@
-// TODO: reuse the code from win32 crate.
-
 use std::{panic::resume_unwind, path::PathBuf};
 
-use windows::{
-    Win32::{
-        Foundation::{ERROR_CANCELLED, HWND},
-        System::Com::{
-            CLSCTX_INPROC_SERVER, COINIT_APARTMENTTHREADED, CoCreateInstance, CoInitializeEx,
-            CoTaskMemFree, CoUninitialize,
-        },
-        UI::Shell::{
-            Common::COMDLG_FILTERSPEC, FOS_ALLOWMULTISELECT, FOS_PICKFOLDERS, FileOpenDialog,
-            FileSaveDialog, IFileDialog, IFileOpenDialog, SIGDN_FILESYSPATH,
-        },
-    },
-    core::{HRESULT, HSTRING, Interface, PCWSTR},
-};
+use widestring::U16CString;
+use windows::Win32::Foundation::HWND;
 use winio_handle::AsWindow;
-use winui3::Microsoft::UI::Xaml as MUX;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct FileFilter {
-    name: HSTRING,
-    pattern: HSTRING,
-}
-
-impl FileFilter {
-    pub fn new(name: &str, pattern: &str) -> Self {
-        Self {
-            name: HSTRING::from(name),
-            pattern: HSTRING::from(pattern),
-        }
-    }
-}
+pub use winio_ui_windows_common::FileFilter;
+use winio_ui_windows_common::filebox;
 
 #[derive(Debug, Default, Clone)]
 pub struct FileBox {
-    title: HSTRING,
-    filename: HSTRING,
+    title: U16CString,
+    filename: U16CString,
     filters: Vec<FileFilter>,
 }
 
@@ -47,11 +19,11 @@ impl FileBox {
     }
 
     pub fn title(&mut self, title: &str) {
-        self.title = HSTRING::from(title);
+        self.title = U16CString::from_str_truncate(title);
     }
 
     pub fn filename(&mut self, filename: &str) {
-        self.filename = HSTRING::from(filename);
+        self.filename = U16CString::from_str_truncate(filename);
     }
 
     pub fn filters(&mut self, filters: impl IntoIterator<Item = FileFilter>) {
@@ -65,6 +37,7 @@ impl FileBox {
     pub async fn open(self, parent: Option<impl AsWindow>) -> Option<PathBuf> {
         let parent = parent.map(|p| p.as_window().as_winui().clone());
         compio::runtime::spawn_blocking(move || unsafe {
+            let parent = parent.and_then(|w| Some(HWND(w.AppWindow().ok()?.Id().ok()?.Value as _)));
             filebox(
                 parent,
                 self.title,
@@ -83,6 +56,7 @@ impl FileBox {
     pub async fn open_multiple(self, parent: Option<impl AsWindow>) -> Vec<PathBuf> {
         let parent = parent.map(|p| p.as_window().as_winui().clone());
         compio::runtime::spawn_blocking(move || unsafe {
+            let parent = parent.and_then(|w| Some(HWND(w.AppWindow().ok()?.Id().ok()?.Value as _)));
             filebox(
                 parent,
                 self.title,
@@ -101,6 +75,7 @@ impl FileBox {
     pub async fn open_folder(self, parent: Option<impl AsWindow>) -> Option<PathBuf> {
         let parent = parent.map(|p| p.as_window().as_winui().clone());
         compio::runtime::spawn_blocking(move || unsafe {
+            let parent = parent.and_then(|w| Some(HWND(w.AppWindow().ok()?.Id().ok()?.Value as _)));
             filebox(
                 parent,
                 self.title,
@@ -119,6 +94,7 @@ impl FileBox {
     pub async fn save(self, parent: Option<impl AsWindow>) -> Option<PathBuf> {
         let parent = parent.map(|p| p.as_window().as_winui().clone());
         compio::runtime::spawn_blocking(move || unsafe {
+            let parent = parent.and_then(|w| Some(HWND(w.AppWindow().ok()?.Id().ok()?.Value as _)));
             filebox(
                 parent,
                 self.title,
@@ -132,125 +108,5 @@ impl FileBox {
         })
         .await
         .unwrap_or_else(|e| resume_unwind(e))
-    }
-}
-
-unsafe fn filebox(
-    parent: Option<MUX::Window>,
-    title: HSTRING,
-    filename: HSTRING,
-    filters: Vec<FileFilter>,
-    open: bool,
-    multiple: bool,
-    folder: bool,
-) -> FileBoxInner {
-    let init = CoInitialize::init();
-
-    let handle: IFileDialog = if open {
-        CoCreateInstance(&FileOpenDialog, None, CLSCTX_INPROC_SERVER).unwrap()
-    } else {
-        CoCreateInstance(&FileSaveDialog, None, CLSCTX_INPROC_SERVER).unwrap()
-    };
-
-    if !title.is_empty() {
-        handle.SetTitle(PCWSTR(title.as_ptr())).unwrap();
-    }
-    if !filename.is_empty() {
-        handle.SetFileName(PCWSTR(filename.as_ptr())).unwrap();
-    }
-
-    let types = filters
-        .iter()
-        .map(|filter| COMDLG_FILTERSPEC {
-            pszName: PCWSTR(filter.name.as_ptr()),
-            pszSpec: PCWSTR(filter.pattern.as_ptr()),
-        })
-        .collect::<Vec<_>>();
-    handle.SetFileTypes(&types).unwrap();
-
-    if multiple {
-        debug_assert!(open, "Cannot save to multiple targets.");
-
-        let mut opts = handle.GetOptions().unwrap();
-        opts |= FOS_ALLOWMULTISELECT;
-        handle.SetOptions(opts).unwrap();
-    }
-
-    if folder {
-        debug_assert!(open, "Cannot save to a folder.");
-
-        let mut opts = handle.GetOptions().unwrap();
-        opts |= FOS_PICKFOLDERS;
-        handle.SetOptions(opts).unwrap();
-    }
-
-    let handle = match handle
-        .Show(parent.and_then(|w| Some(HWND(w.AppWindow().ok()?.Id().ok()?.Value as _))))
-    {
-        Ok(()) => Some(handle),
-        Err(e) if e.code() == HRESULT::from(ERROR_CANCELLED) => None,
-        Err(e) => panic!("{e:?}"),
-    };
-
-    FileBoxInner(handle, init)
-}
-
-struct FileBoxInner(Option<IFileDialog>, CoInitialize);
-
-impl FileBoxInner {
-    pub unsafe fn result(self) -> Option<PathBuf> {
-        if let Some(dialog) = self.0 {
-            let item = dialog.GetResult().unwrap();
-            let name_ptr = item.GetDisplayName(SIGDN_FILESYSPATH).unwrap();
-            let name_ptr = CoTaskMemPtr(name_ptr.0);
-            let name = PCWSTR::from_raw(name_ptr.0).to_hstring().to_os_string();
-            Some(PathBuf::from(name))
-        } else {
-            None
-        }
-    }
-
-    pub unsafe fn results(self) -> Vec<PathBuf> {
-        if let Some(dialog) = self.0 {
-            let handle: IFileOpenDialog = dialog.cast().unwrap();
-            let results = handle.GetResults().unwrap();
-            let count = results.GetCount().unwrap();
-            let mut names = vec![];
-            for i in 0..count {
-                let item = results.GetItemAt(i).unwrap();
-                let name_ptr = item.GetDisplayName(SIGDN_FILESYSPATH).unwrap();
-                let name_ptr = CoTaskMemPtr(name_ptr.0);
-                let name = PCWSTR::from_raw(name_ptr.0).to_hstring().to_os_string();
-                names.push(PathBuf::from(name));
-            }
-            names
-        } else {
-            vec![]
-        }
-    }
-}
-
-struct CoTaskMemPtr<T>(*mut T);
-
-impl<T> Drop for CoTaskMemPtr<T> {
-    fn drop(&mut self) {
-        unsafe { CoTaskMemFree(Some(self.0.cast())) }
-    }
-}
-
-struct CoInitialize;
-
-impl CoInitialize {
-    pub fn init() -> Self {
-        unsafe {
-            CoInitializeEx(None, COINIT_APARTMENTTHREADED).unwrap();
-        }
-        Self
-    }
-}
-
-impl Drop for CoInitialize {
-    fn drop(&mut self) {
-        unsafe { CoUninitialize() };
     }
 }
