+++ winio-ui-winui/src/ui/mod.rs
@@ -4,6 +4,7 @@ use windows::{
     core::{HSTRING, Interface, RuntimeType},
 };
 use winio_primitive::{ColorTheme, HAlign, Point, Size};
+pub use winio_ui_windows_common::{accent_color, monitor_get_all};
 use winui3::Microsoft::UI::Xaml::{Application, TextAlignment};
 
 trait Convertible<T> {
@@ -129,6 +130,3 @@ pub use filebox::*;
 
 mod msgbox;
 pub use msgbox::*;
-
-mod accent;
-pub use accent::*;
